Reorganized project structure by replacing "models" with "entity" and "dto" folder.
- Deleted "models" folder.
- Created "entity" and "dto" foler to separate database entities and request models.
