+++ internal/v1/user/entity/user.go
@@ -1,0 +1,9 @@
+package entity
+
+type User struct {
+	ID       uint   `gorm:"primaryKey"`
+	Username string `gorm:"unique;not null"`
+	Email    string `gorm:"unique;not null"`
+	Password string `gorm:"not null"`
+	Role     string `gorm:"not null"`
+}
