--- internal/v1/user/models/auth.go
@@ -1,1 +1,0 @@
-package models
