+++ internal/v1/user/repository/user_repository.go
@@ -1,16 +1,16 @@
 package repository
 
 import (
-	"eccomerce/internal/v1/user/models"
+	"eccomerce/internal/v1/user/entity"
 
 	"gorm.io/gorm"
 )
 
 type UserRepository interface {
-	Create(user *models.User) error
-	GetByID(id uint) (*models.User, error)
-	GetByEmail(email string) (*models.User, error)
-	Update(user *models.User) error
+	Create(user *entity.User) error
+	GetByID(id uint) (*entity.User, error)
+	GetByEmail(email string) (*entity.User, error)
+	Update(user *entity.User) error
 	Delete(id uint) error
 }
 
@@ -22,30 +22,30 @@ func NewUserRepository(db *gorm.DB) UserRepository {
 	return &userRepository{db: db}
 }
 
-func (r *userRepository) Create(user *models.User) error {
+func (r *userRepository) Create(user *entity.User) error {
 	return r.db.Create(user).Error
 }
 
-func (r *userRepository) GetByID(id uint) (*models.User, error) {
-	var user models.User
+func (r *userRepository) GetByID(id uint) (*entity.User, error) {
+	var user entity.User
 	if err := r.db.First(&user, id).Error; err != nil {
 		return nil, err
 	}
 	return &user, nil
 }
 
-func (r *userRepository) GetByEmail(email string) (*models.User, error) {
-	var user models.User
+func (r *userRepository) GetByEmail(email string) (*entity.User, error) {
+	var user entity.User
 	if err := r.db.Where("email = ?", email).First(&user).Error; err != nil {
 		return nil, err
 	}
 	return &user, nil
 }
 
-func (r *userRepository) Update(user *models.User) error {
+func (r *userRepository) Update(user *entity.User) error {
 	return r.db.Save(user).Error
 }
 
 func (r *userRepository) Delete(id uint) error {
-	return r.db.Delete(&models.User{}, id).Error
+	return r.db.Delete(&entity.User{}, id).Error
 }
